feature > implement ajax when adding to cart
Product page. Replace “Add to cart” button with a JavaScript ajax request “Add to cart” that should add the product to the cart with a Shopify line item property of key “cart_index” and value the index number in the cart. For example, if there are no items in the cart, the value should be 0, if there are already 3 products in the cart, the index should be 3.
“Add to cart” button should be grey and disabled when there is no stock. When there is stock, it should be green and blue on hover.