add textarea for order notes
In the Cart page, add a checkbox that says “Add instructions”, when enabled/selected, show a textarea that will let the customer add text. When the customer completes the order, this text should be in the Order Notes. If the checkbox is unselected/disabled, no text should appear in the Order Notes.