add news carousel to homepage
Create a new Section named “news-carousel.liquid”, where 3 different types of blocks of information will be shown, configured with the Shopify CMS. The section has to be in the home page.
There will be 3 different blocks: campaign, new-product, news. All of these blocks will have the same attributes/settings:
Title (text)
Description (textarea)
Featured (checkbox)
The section has to render all the different blocks, showing the title and description. The title for the campaign has to be red, the title for the new-product has to be blue and the title for the news has to be green. Blocks who have the “Featured” attribute enabled/selected, have to appear first in the list.