fr > implement a script to create/update metafield
Using a Shopify Private API Key, that can be generated in the development store, create a script in Node that will create a product metafield (namespace: global, key: test, type: integer), if the metafield does not exists, the metafield value should be 0, if the metafield already exists, add +1 to the value. The script is executed manually.